% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{boot_db_connect}
\alias{boot_db_connect}
\title{Connect to Simulation Database}
\usage{
boot_db_connect(path = "simulations.duckdb", read_only = FALSE)
}
\arguments{
\item{path}{Character. Path to the DuckDB file. Default is "simulations.duckdb".}

\item{read_only}{Logical. If TRUE, open in read-only mode. Default is FALSE.}
}
\value{
A DBI connection object.
}
\description{
Opens a connection to a DuckDB database file. Creates the file if it
doesn't exist, appends if it does.
}
\details{
Requires the \code{duckdb} package to be installed. The database file is created
if it doesn't exist. Use \code{\link[=boot_db_init]{boot_db_init()}} after connecting to create the
schema if this is a new database.
}
\examples{
\dontrun{
con <- boot_db_connect("my_simulations.duckdb")
boot_db_init(con)
# ... run simulations ...
DBI::dbDisconnect(con)
}

}
\seealso{
\code{\link[=boot_db_init]{boot_db_init()}}, \code{\link[=boot_db_write]{boot_db_write()}}
}
