% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{burg_aic_select}
\alias{burg_aic_select}
\title{Burg AR with AIC Selection (C++ Implementation)}
\usage{
burg_aic_select(x, maxp = 5, criterion = c("aic", "aicc", "bic"))
}
\arguments{
\item{x}{Numeric vector, the time series (should be centered).}

\item{maxp}{Integer, maximum AR order to consider.}

\item{criterion}{Character, information criterion: "aic", "aicc", or "bic".}
}
\value{
List with components:
\item{p}{Selected AR order.}
\item{phi}{AR coefficients.}
\item{vara}{Innovation variance.}
}
\description{
Select AR order using information criterion and fit coefficients.
}
