% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{burg_fit}
\alias{burg_fit}
\title{Burg AR Fit (C++ Implementation)}
\usage{
burg_fit(x, p)
}
\arguments{
\item{x}{Numeric vector, the time series.}

\item{p}{Integer, AR order to fit.}
}
\value{
Numeric vector of AR coefficients.
}
\description{
Fit AR coefficients using Burg's algorithm.
}
