% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{co_tstat}
\alias{co_tstat}
\title{Cochrane-Orcutt t-statistic (C++ Implementation)}
\usage{
co_tstat(x, maxp = 5, criterion = c("aic", "aicc", "bic"))
}
\arguments{
\item{x}{Numeric vector, the time series.}

\item{maxp}{Integer, maximum AR order for model selection.}

\item{criterion}{Character, information criterion: "aic", "aicc", or "bic".}
}
\value{
Numeric, the t-statistic for trend.
}
\description{
Compute the CO t-statistic for testing trend.
}
