% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovation_generators.R
\name{innovation_generators}
\alias{innovation_generators}
\title{Innovation Generators}
\description{
Factory functions for creating innovation generators for time series simulation.
}
\keyword{internal}
