% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{mc_db_connect}
\alias{mc_db_connect}
\title{Connect to Simulation Database}
\usage{
mc_db_connect(path = "simulations.duckdb", read_only = FALSE)
}
\arguments{
\item{path}{Character. Path to the DuckDB file. Default is "simulations.duckdb".}

\item{read_only}{Logical. If TRUE, open in read-only mode. Default is FALSE.}
}
\value{
A DBI connection object.
}
\description{
Opens a connection to a DuckDB database file. Creates the file if it
doesn't exist, appends if it does.
}
\examples{
\dontrun{
con <- mc_db_connect("capstone.duckdb")
mc_db_init(con)
DBI::dbDisconnect(con)
}

}
\seealso{
\code{\link[=mc_db_init]{mc_db_init()}}
}
