% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstse-package.R
\docType{package}
\name{tstse-package}
\alias{tstse}
\alias{tstse-package}
\title{tstse: Time Series for Data Science}
\description{
An experimental rewrite of the tswge package. Accompanies the texts Time Series for Data Science with R by Woodward, Sadler and Robertson & Applied Time Series Analysis with R, 2nd edition by Woodward, Gray, and Elliott. It is helpful for data analysis and for time series instruction. Includes tools for GARCH model comparison and diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Aix3lsyD/tstse}
  \item Report bugs at \url{https://github.com/Aix3lsyD/tstse/issues}
}

}
\author{
\strong{Maintainer}: Bivin Sadler \email{bsadler@smu.edu}

Authors:
\itemize{
  \item Wayne Woodward
}

Other contributors:
\itemize{
  \item Kyle Evans [contributor]
}

}
\keyword{internal}
